cvbuilder/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
httpx==0.28.1
crossref-commons==0.0.7
weasyprint==63.1
jinja2==3.1.4
pyyaml==6.0.2
aiofiles==24.1.0
